+++ src/test/java/org/folio/inventory/dataimport/consumers/vertices/DataImportConsumerVerticleTest.java
@@ -1,4 +1,4 @@
-package org.folio.inventory.dataimport.consumers.verticles;
+package org.folio.inventory.dataimport.consumers.vertices;
 
 import static com.github.tomakehurst.wiremock.client.WireMock.get;
 import static java.nio.charset.StandardCharsets.UTF_8;
