+++ src/test/java/org/folio/inventory/dataimport/consumers/vertices/MarcBibUpdateConsumerVerticleTest.java
@@ -1,4 +1,12 @@
-package org.folio.inventory.dataimport.consumers.verticles;
+package org.folio.inventory.dataimport.consumers.vertices;
+
+import static net.mguenther.kafka.junit.EmbeddedKafkaCluster.provisionWith;
+import static net.mguenther.kafka.junit.EmbeddedKafkaClusterConfig.defaultClusterConfig;
+import static org.folio.inventory.dataimport.util.KafkaConfigConstants.KAFKA_ENV;
+import static org.folio.inventory.dataimport.util.KafkaConfigConstants.KAFKA_HOST;
+import static org.folio.inventory.dataimport.util.KafkaConfigConstants.KAFKA_MAX_REQUEST_SIZE;
+import static org.folio.inventory.dataimport.util.KafkaConfigConstants.KAFKA_PORT;
+import static org.folio.inventory.dataimport.util.KafkaConfigConstants.KAFKA_REPLICATION_FACTOR;
 
 import io.vertx.core.DeploymentOptions;
 import io.vertx.core.Promise;
@@ -8,22 +16,28 @@ import io.vertx.ext.unit.Async;
 import io.vertx.ext.unit.TestContext;
 import io.vertx.ext.unit.junit.VertxUnitRunner;
 import net.mguenther.kafka.junit.EmbeddedKafkaCluster;
-import org.folio.inventory.InstanceIngressConsumerVerticle;
+import org.folio.inventory.MarcBibUpdateConsumerVerticle;
+import org.folio.inventory.dataimport.consumers.MarcBibUpdateKafkaHandler;
 import org.junit.AfterClass;
+import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-
-import static net.mguenther.kafka.junit.EmbeddedKafkaCluster.provisionWith;
-import static net.mguenther.kafka.junit.EmbeddedKafkaClusterConfig.defaultClusterConfig;
-import static org.folio.inventory.dataimport.util.KafkaConfigConstants.*;
+import org.mockito.Mock;
+import org.mockito.MockitoAnnotations;
 
 @RunWith(VertxUnitRunner.class)
-public class InstanceIngressConsumerVerticleTest {
-
+public class MarcBibUpdateConsumerVerticleTest {
   private static final String KAFKA_ENV_NAME = "test-env";
   private static Vertx vertx = Vertx.vertx();
-
   public static EmbeddedKafkaCluster cluster;
+
+  @Mock
+  private static MarcBibUpdateKafkaHandler marcBibUpdateKafkaHandler;
+
+  @Before
+  public void setUp() {
+    MockitoAnnotations.openMocks(this);
+  }
 
   @Test
   public void shouldDeployVerticle(TestContext context) {
@@ -41,7 +55,7 @@ public class InstanceIngressConsumerVerticleTest {
       .setWorker(true);
 
     Promise<String> promise = Promise.promise();
-    vertx.deployVerticle(InstanceIngressConsumerVerticle.class.getName(), options, promise);
+    vertx.deployVerticle(MarcBibUpdateConsumerVerticle.class.getName(), options, promise);
 
     promise.future().onComplete(ar -> {
       context.assertTrue(ar.succeeded());
